+++ src/base_formatter/context.rs
@@ -1,0 +1,627 @@
+use std::fmt;
+use std::fmt::{Debug, Display};
+use std::num::ParseIntError;
+use std::str::FromStr;
+
+use crate::base_formatter::printer::PrinterOptions;
+use crate::base_formatter::token::string::Quote;
+
+/// Context object storing data relevant when formatting an object.
+pub trait FormatContext {
+    type Options: FormatOptions;
+
+    /// Returns the formatting options
+    fn options(&self) -> &Self::Options;
+}
+
+/// Options customizing how the source code should be formatted.
+///
+/// **Note**: This trait should **only** contain the essential abstractions required for the printing phase.
+/// For example, do not add a `fn bracket_spacing(&self) -> BracketSpacing` method here,
+/// as the [BracketSpacing] option is not needed during the printing phase
+/// and enforcing its implementation for all structs using this trait is unnecessary.
+pub trait FormatOptions {
+    /// The indent style.
+    fn indent_style(&self) -> IndentStyle;
+
+    /// The indent width.
+    fn indent_width(&self) -> IndentWidth;
+
+    /// What's the max width of a line. Defaults to 80.
+    fn line_width(&self) -> LineWidth;
+
+    /// The type of line ending.
+    fn line_ending(&self) -> LineEnding;
+
+    /// Derives the print options from the these format options
+    fn as_print_options(&self) -> PrinterOptions;
+}
+
+// ---
+
+// NOTE: Only used for tests now
+#[derive(Debug, Default, Eq, PartialEq)]
+pub struct SimpleFormatContext {
+    options: SimpleFormatOptions,
+}
+
+impl SimpleFormatContext {
+    pub fn new(options: SimpleFormatOptions) -> Self {
+        Self { options }
+    }
+}
+
+impl FormatContext for SimpleFormatContext {
+    type Options = SimpleFormatOptions;
+
+    fn options(&self) -> &Self::Options {
+        &self.options
+    }
+}
+
+// NOTE: Only used for tests now
+#[derive(Debug, Default, Eq, PartialEq, Copy, Clone)]
+pub struct SimpleFormatOptions {
+    pub indent_style: IndentStyle,
+    pub indent_width: IndentWidth,
+    pub line_width: LineWidth,
+    pub line_ending: LineEnding,
+}
+
+impl FormatOptions for SimpleFormatOptions {
+    fn indent_style(&self) -> IndentStyle {
+        self.indent_style
+    }
+
+    fn indent_width(&self) -> IndentWidth {
+        self.indent_width
+    }
+
+    fn line_width(&self) -> LineWidth {
+        self.line_width
+    }
+
+    fn line_ending(&self) -> LineEnding {
+        self.line_ending
+    }
+
+    fn as_print_options(&self) -> PrinterOptions {
+        PrinterOptions::default()
+            .with_indent_style(self.indent_style)
+            .with_indent_width(self.indent_width)
+            .with_print_width(self.line_width.into())
+            .with_line_ending(self.line_ending)
+    }
+}
+
+impl Display for SimpleFormatOptions {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        fmt::Debug::fmt(self, f)
+    }
+}
+
+// ---
+
+#[derive(Debug, Default, Clone, Copy, Eq, Hash, PartialEq)]
+pub enum IndentStyle {
+    /// Tab
+    #[default]
+    Tab,
+    /// Space
+    Space,
+}
+
+impl IndentStyle {
+    pub const DEFAULT_SPACES: u8 = 2;
+
+    /// Returns `true` if this is an [IndentStyle::Tab].
+    pub const fn is_tab(&self) -> bool {
+        matches!(self, IndentStyle::Tab)
+    }
+
+    /// Returns `true` if this is an [IndentStyle::Space].
+    pub const fn is_space(&self) -> bool {
+        matches!(self, IndentStyle::Space)
+    }
+}
+
+impl FromStr for IndentStyle {
+    type Err = &'static str;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        match s {
+            "tab" => Ok(Self::Tab),
+            "space" => Ok(Self::Space),
+            // TODO: replace this error with a diagnostic
+            _ => Err("Unsupported value for this option"),
+        }
+    }
+}
+
+impl Display for IndentStyle {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            IndentStyle::Tab => std::write!(f, "Tab"),
+            IndentStyle::Space => std::write!(f, "Space"),
+        }
+    }
+}
+
+#[derive(Clone, Copy, Debug, Eq, Hash, PartialEq, Default)]
+pub enum LineEnding {
+    ///  Line Feed only (\n), common on Linux and macOS as well as inside git repos
+    #[default]
+    Lf,
+
+    /// Carriage Return + Line Feed characters (\r\n), common on Windows
+    Crlf,
+
+    /// Carriage Return character only (\r), used very rarely
+    Cr,
+}
+
+impl LineEnding {
+    #[inline]
+    pub const fn as_str(&self) -> &'static str {
+        match self {
+            LineEnding::Lf => "\n",
+            LineEnding::Crlf => "\r\n",
+            LineEnding::Cr => "\r",
+        }
+    }
+
+    /// Returns `true` if this is a [LineEnding::Lf].
+    pub const fn is_line_feed(&self) -> bool {
+        matches!(self, LineEnding::Lf)
+    }
+
+    /// Returns `true` if this is a [LineEnding::Crlf].
+    pub const fn is_carriage_return_line_feed(&self) -> bool {
+        matches!(self, LineEnding::Crlf)
+    }
+
+    /// Returns `true` if this is a [LineEnding::Cr].
+    pub const fn is_carriage_return(&self) -> bool {
+        matches!(self, LineEnding::Cr)
+    }
+}
+
+impl FromStr for LineEnding {
+    type Err = &'static str;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        match s {
+            "lf" => Ok(Self::Lf),
+            "crlf" => Ok(Self::Crlf),
+            "cr" => Ok(Self::Cr),
+            // TODO: replace this error with a diagnostic
+            _ => Err("Value not supported for LineEnding"),
+        }
+    }
+}
+
+impl std::fmt::Display for LineEnding {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            LineEnding::Lf => std::write!(f, "LF"),
+            LineEnding::Crlf => std::write!(f, "CRLF"),
+            LineEnding::Cr => std::write!(f, "CR"),
+        }
+    }
+}
+
+#[derive(Clone, Copy, Eq, Hash, PartialEq)]
+pub struct IndentWidth(u8);
+
+impl IndentWidth {
+    pub const MIN: u8 = 0;
+
+    pub const MAX: u8 = 24;
+
+    /// Return the numeric value for this [IndentWidth]
+    pub fn value(&self) -> u8 {
+        self.0
+    }
+}
+
+impl Default for IndentWidth {
+    fn default() -> Self {
+        Self(2)
+    }
+}
+
+impl FromStr for IndentWidth {
+    type Err = ParseFormatNumberError;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        let value = u8::from_str(s).map_err(ParseFormatNumberError::ParseError)?;
+        let value = Self::try_from(value).map_err(ParseFormatNumberError::TryFromU8Error)?;
+        Ok(value)
+    }
+}
+
+impl TryFrom<u8> for IndentWidth {
+    type Error = IndentWidthFromIntError;
+
+    fn try_from(value: u8) -> Result<Self, Self::Error> {
+        if (Self::MIN..=Self::MAX).contains(&value) {
+            Ok(Self(value))
+        } else {
+            Err(IndentWidthFromIntError(value))
+        }
+    }
+}
+
+impl Display for IndentWidth {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        let value = self.value();
+        f.write_str(&std::format!("{value}"))
+    }
+}
+
+impl Debug for IndentWidth {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        std::fmt::Display::fmt(self, f)
+    }
+}
+
+/// Validated value for the `line_width` formatter options
+///
+/// The allowed range of values is 1..=320
+#[derive(Clone, Copy, Eq, PartialEq)]
+pub struct LineWidth(u16);
+
+impl LineWidth {
+    /// Minimum allowed value for a valid [LineWidth]
+    pub const MIN: u16 = 1;
+    /// Maximum allowed value for a valid [LineWidth]
+    pub const MAX: u16 = 320;
+
+    /// Return the numeric value for this [LineWidth]
+    pub fn value(&self) -> u16 {
+        self.0
+    }
+}
+
+impl Default for LineWidth {
+    fn default() -> Self {
+        Self(80)
+    }
+}
+
+impl Display for LineWidth {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        let value = self.value();
+        f.write_str(&std::format!("{value}"))
+    }
+}
+
+impl Debug for LineWidth {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        std::fmt::Display::fmt(self, f)
+    }
+}
+
+/// Error type returned when parsing a [LineWidth] or [IndentWidth] from a string fails
+pub enum ParseFormatNumberError {
+    /// The string could not be parsed to a number
+    ParseError(ParseIntError),
+    /// The `u16` value of the string is not a valid [LineWidth]
+    TryFromU16Error(LineWidthFromIntError),
+    /// The `u8 value of the string is not a valid [IndentWidth]
+    TryFromU8Error(IndentWidthFromIntError),
+}
+
+impl From<IndentWidthFromIntError> for ParseFormatNumberError {
+    fn from(value: IndentWidthFromIntError) -> Self {
+        Self::TryFromU8Error(value)
+    }
+}
+
+impl From<LineWidthFromIntError> for ParseFormatNumberError {
+    fn from(value: LineWidthFromIntError) -> Self {
+        Self::TryFromU16Error(value)
+    }
+}
+
+impl From<ParseIntError> for ParseFormatNumberError {
+    fn from(value: ParseIntError) -> Self {
+        Self::ParseError(value)
+    }
+}
+
+impl Debug for ParseFormatNumberError {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        std::fmt::Display::fmt(self, f)
+    }
+}
+
+impl std::fmt::Display for ParseFormatNumberError {
+    fn fmt(&self, fmt: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            ParseFormatNumberError::ParseError(err) => std::fmt::Display::fmt(err, fmt),
+            ParseFormatNumberError::TryFromU16Error(err) => std::fmt::Display::fmt(err, fmt),
+            ParseFormatNumberError::TryFromU8Error(err) => std::fmt::Display::fmt(err, fmt),
+        }
+    }
+}
+
+impl TryFrom<u16> for LineWidth {
+    type Error = LineWidthFromIntError;
+
+    fn try_from(value: u16) -> Result<Self, Self::Error> {
+        if (Self::MIN..=Self::MAX).contains(&value) {
+            Ok(Self(value))
+        } else {
+            Err(LineWidthFromIntError(value))
+        }
+    }
+}
+
+impl FromStr for LineWidth {
+    type Err = ParseFormatNumberError;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        let value = u16::from_str(s).map_err(ParseFormatNumberError::ParseError)?;
+        let value = Self::try_from(value).map_err(ParseFormatNumberError::TryFromU16Error)?;
+        Ok(value)
+    }
+}
+
+/// Error type returned when converting a u16 to a [LineWidth] fails
+#[derive(Clone, Copy, Debug)]
+pub struct IndentWidthFromIntError(pub u8);
+
+impl std::fmt::Display for IndentWidthFromIntError {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        writeln!(
+            f,
+            "The indent width should be between {} and {}",
+            LineWidth::MIN,
+            LineWidth::MAX,
+        )
+    }
+}
+
+/// Error type returned when converting a u16 to a [LineWidth] fails
+#[derive(Clone, Copy, Debug)]
+pub struct LineWidthFromIntError(pub u16);
+
+impl std::fmt::Display for LineWidthFromIntError {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        writeln!(
+            f,
+            "The line width should be between {} and {}",
+            LineWidth::MIN,
+            LineWidth::MAX,
+        )
+    }
+}
+
+impl From<LineWidth> for u16 {
+    fn from(value: LineWidth) -> Self {
+        value.0
+    }
+}
+
+#[derive(Clone, Copy, Debug, Default, Eq, Hash, PartialEq)]
+pub enum QuoteStyle {
+    #[default]
+    Double,
+    Single,
+}
+
+impl QuoteStyle {
+    pub fn from_byte(byte: u8) -> Option<QuoteStyle> {
+        match byte {
+            b'"' => Some(QuoteStyle::Double),
+            b'\'' => Some(QuoteStyle::Single),
+            _ => None,
+        }
+    }
+
+    pub fn as_char(&self) -> char {
+        match self {
+            QuoteStyle::Double => '"',
+            QuoteStyle::Single => '\'',
+        }
+    }
+
+    pub fn as_byte(&self) -> u8 {
+        self.as_char() as u8
+    }
+
+    /// Returns the quote in HTML entity
+    pub fn as_html_entity(&self) -> &str {
+        match self {
+            QuoteStyle::Double => "&quot;",
+            QuoteStyle::Single => "&apos;",
+        }
+    }
+
+    /// Given the current quote, it returns the other one
+    pub fn other(&self) -> Self {
+        match self {
+            QuoteStyle::Double => QuoteStyle::Single,
+            QuoteStyle::Single => QuoteStyle::Double,
+        }
+    }
+
+    pub const fn is_double(&self) -> bool {
+        matches!(self, Self::Double)
+    }
+}
+
+impl FromStr for QuoteStyle {
+    type Err = &'static str;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        match s {
+            "double" => Ok(Self::Double),
+            "single" => Ok(Self::Single),
+            // TODO: replace this error with a diagnostic
+            _ => Err("Value not supported for QuoteStyle"),
+        }
+    }
+}
+
+impl std::fmt::Display for QuoteStyle {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            QuoteStyle::Double => std::write!(f, "Double Quotes"),
+            QuoteStyle::Single => std::write!(f, "Single Quotes"),
+        }
+    }
+}
+
+impl From<QuoteStyle> for Quote {
+    fn from(quote: QuoteStyle) -> Self {
+        match quote {
+            QuoteStyle::Double => Self::Double,
+            QuoteStyle::Single => Self::Single,
+        }
+    }
+}
+
+#[derive(Clone, Copy, Debug, Eq, Hash, PartialEq)]
+pub struct BracketSpacing(bool);
+
+impl BracketSpacing {
+    /// Return the boolean value for this [BracketSpacing]
+    pub fn value(&self) -> bool {
+        self.0
+    }
+}
+
+impl Default for BracketSpacing {
+    fn default() -> Self {
+        Self(true)
+    }
+}
+
+impl From<bool> for BracketSpacing {
+    fn from(value: bool) -> Self {
+        Self(value)
+    }
+}
+
+impl std::fmt::Display for BracketSpacing {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        std::write!(f, "{:?}", self.value())
+    }
+}
+
+impl FromStr for BracketSpacing {
+    type Err = &'static str;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        let value = bool::from_str(s);
+
+        match value {
+            Ok(value) => Ok(Self(value)),
+            Err(_) => Err(
+                "Value not supported for BracketSpacing. Supported values are 'true' and 'false'.",
+            ),
+        }
+    }
+}
+
+#[derive(Clone, Copy, Debug, Default, Eq, Hash, PartialEq)]
+pub enum AttributePosition {
+    #[default]
+    Auto,
+    Multiline,
+}
+
+impl std::fmt::Display for AttributePosition {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            AttributePosition::Auto => std::write!(f, "Auto"),
+            AttributePosition::Multiline => std::write!(f, "Multiline"),
+        }
+    }
+}
+
+impl FromStr for AttributePosition {
+    type Err = &'static str;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        match s {
+            "multiline" => Ok(Self::Multiline),
+            "auto" => Ok(Self::Auto),
+            _ => Err(
+                "Value not supported for attribute_position. Supported values are 'auto' and 'multiline'.",
+            ),
+        }
+    }
+}
+
+/// Put the `>` of a multi-line HTML or JSX element at the end of the last line instead of being alone on the next line (does not apply to self closing elements).
+#[derive(Clone, Copy, Debug, Default, Eq, Hash, PartialEq)]
+pub struct BracketSameLine(bool);
+
+impl BracketSameLine {
+    /// Return the boolean value for this [BracketSameLine]
+    pub fn value(&self) -> bool {
+        self.0
+    }
+}
+
+impl From<bool> for BracketSameLine {
+    fn from(value: bool) -> Self {
+        Self(value)
+    }
+}
+
+impl std::fmt::Display for BracketSameLine {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        std::write!(f, "{}", self.value())
+    }
+}
+
+impl FromStr for BracketSameLine {
+    type Err = &'static str;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        match bool::from_str(s) {
+            Ok(value) => Ok(Self(value)),
+            Err(_) => Err(
+                "Value not supported for BracketSameLine. Supported values are 'true' and 'false'.",
+            ),
+        }
+    }
+}
+
+#[derive(Clone, Copy, Debug, Default, Eq, Hash, PartialEq)]
+pub enum Expand {
+    /// Objects are expanded when the first property has a leading newline. Arrays are always
+    /// expanded if they are shorter than the line width.
+    #[default]
+    Auto,
+    /// Objects and arrays are always expanded.
+    Always,
+    /// Objects and arrays are never expanded, if they are shorter than the line width.
+    Never,
+}
+
+impl FromStr for Expand {
+    type Err = String;
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        match s {
+            "auto" => Ok(Self::Auto),
+            "always" => Ok(Self::Always),
+            "never" => Ok(Self::Never),
+            _ => Err(std::format!("unknown expand literal: {}", s)),
+        }
+    }
+}
+
+impl fmt::Display for Expand {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        match self {
+            Expand::Auto => std::write!(f, "Auto"),
+            Expand::Always => std::write!(f, "Always"),
+            Expand::Never => std::write!(f, "Never"),
+        }
+    }
+}
